// src/enums/operation-type.enum.ts
export enum OperationTypeEnum {
  addRowInTable = 'addRowInTable',
  deleteRowInTable = 'deleteRowInTable',
  getRowByPrimaryKey = 'getRowByPrimaryKey',
  getRowsFromTable = 'getRowsFromTable',
  getTableForeignKeys = 'getTableForeignKeys',
  getTablePrimaryColumns = 'getTablePrimaryColumns',
  getTableStructure = 'getTableStructure',
  getTablesFromDB = 'getTablesFromDB',
  testConnect = 'testConnect',
  updateRowInTable = 'updateRowInTable',
  validateSettings = 'validateSettings',
  initialConnection = 'initialConnection',
  dataFromAgent = 'dataFromAgent',
}